Merge 120715 - Early-return in RenderViewImpl::DidFlushPaint() if there is no WebView.
CL 9328016 exposed a bug with seemingly-unused stats code assuming that the RenderView's WebView must be valid when DidFlushPaint() is called, which may not be true if RenderWidget::OnSwapBuffersComplete is notified while the widget is closing down.

We'll follow up by verifying the stats code is unnecessary and removing it (see bug 112942).

BUG=112921,111639,112942


Review URL: http://codereview.chromium.org/9349004

TBR=wez@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9363030

git-svn-id: http://src.chromium.org/svn/branches/1025/src@121039 4ff67af0-8c30-449e-8e8b-ad334ec8d88c