Merge 121153 - Ensure that we record sync promo suppressions
Currently we record which version of the sync promo a user was shown from the SyncPromoUI constructor.

This lead to a problem where we didn't count the "don't show sync promo" trial. To fix this we explicitly record the "don't show" trial in BrowserInit.

BUG=113135
TEST=


Review URL: http://codereview.chromium.org/9358017

TBR=sail@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9371023

git-svn-id: http://src.chromium.org/svn/branches/1025/src@121280 4ff67af0-8c30-449e-8e8b-ad334ec8d88c