Merge 120916 - Avoid overread in RenderWidgetHostViewWin::OnPaint
GetRegionData may return 0 to indicate error. This call is used both to
calculate the required buffer size and to fill the buffer with region
data. In the first case, if not checked, a zero-length buffer may be
allocated. In the second case, the buffer content is undefined. In either
case, we should not depend on the buffer content.

BUG=110176
TEST=Long tail crasher in gfx::Rect::Rect, called from
RenderWidgetHostViewWin::OnPaint.


Review URL: http://codereview.chromium.org/9316056

TBR=davidbarr@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9371031

git-svn-id: http://src.chromium.org/svn/branches/1025/src@121346 4ff67af0-8c30-449e-8e8b-ad334ec8d88c